# [SPARK-19088][SQL] Optimize sequence type deserialization codegen