--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/objects/objects.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/objects/objects.scala
@@ -728,7 +728,6 @@ case class CollectObjects private(
 
       if (!${genInputData.isNull}) {
         $determineCollectionType
-        $convertedType[] $convertedArray = null;
         int $dataLength = $getLength;
         ${classOf[Builder[_, _]].getName} $builderValue = $getBuilderVar;
         $builderValue.sizeHint($dataLength);