@rxin 2017-01-11
Is this a perf optimization? If yes, can you show some benchmarks? Also for codegen it's good to show the generated code before/after this change. You can get that with

```
df.queryExecution.debug.codegen()
```