@kiszk 2017-01-16
It looks like the similar optimization to https://github.com/apache/spark/pull/15044. 
Does [this code](https://github.com/apache/spark/pull/15044/files#diff-d6f03c9d3e82f3774d1110559b039a6d) help you?