@kiszk 2017-01-16
Can we get additional performance improvement if we could generate `UnsafeArrayData` instead of `GenericArrayData` for this statement ```/* 104 */       final ArrayData serializefromobject_value = false ? null : new org.apache.spark.sql.catalyst.util.GenericArrayData(serializefromobject_argValue);```?