@michalsenkyr 2017-01-19
I added the benchmarks based on the code you provided but I am getting almost the same results before and after the optimization (see description). So either the added benefit is really small or I didn't write/tune the benchmarks quite right. I would appreciate if you could take a look at them.
I will also take a look at the `UnsafeArrayData` optimization later and try to include it in my next commit.